requirements.txt
fritzconnection
requests
black
pylama
bandit
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
python-dotenv